requirements/common.txt
alembic
apscheduler
coloredlogs
dnspython
flask-restx
isodate
music_metadata_extractor
psycopg2
psycopg2-binary
pymongo
python-dotenv
requests
marshmallow
uwsgi